backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
Werkzeug==2.3.7
pg8000==1.30.3
python-dotenv==1.0.0
Flask-Migrate==4.0.5
